Move `getContactIdsByGroupMembership` to AndroidAddressBook; make LocalAddressBook methods suspendable
